+++ src/service/transaction/pay.ts
@@ -12,6 +12,7 @@ import { MongoRepository as EventRepo } from '../../repo/event';
 import { MongoRepository as ProductRepo } from '../../repo/product';
 import { MongoRepository as ProjectRepo } from '../../repo/project';
 import { MongoRepository as SellerRepo } from '../../repo/seller';
+import { MongoRepository as ServiceOutputRepo } from '../../repo/serviceOutput';
 import { MongoRepository as TaskRepo } from '../../repo/task';
 
 import * as AccountPayment from '../payment/account';
@@ -28,6 +29,7 @@ export type IStartOperation<T> = (repos: {
     product: ProductRepo;
     project: ProjectRepo;
     seller: SellerRepo;
+    serviceOutput: ServiceOutputRepo;
     transaction: TransactionRepo;
     task: TaskRepo;
 }) => Promise<T>;
@@ -52,6 +54,67 @@ export type ICheckOperation<T> = (repos: {
     project: ProjectRepo;
     seller: SellerRepo;
 }) => Promise<T>;
+
+export type IPublishPaymentUrlOperation<T> = (repos: {
+    product: ProductRepo;
+    project: ProjectRepo;
+    seller: SellerRepo;
+}) => Promise<T>;
+
+export interface IPublishPaymentUrlResult {
+    paymentMethodId: string;
+    paymentUrl: string;
+}
+
+/**
+ * 外部決済ロケーション発行
+ */
+export function publishPaymentUrl(
+    params: factory.assetTransaction.pay.IStartParamsWithoutDetail
+): IPublishPaymentUrlOperation<IPublishPaymentUrlResult> {
+    return async (repos: {
+        product: ProductRepo;
+        project: ProjectRepo;
+        seller: SellerRepo;
+    }) => {
+        const paymentServiceType = params.object?.typeOf;
+
+        // 金額をfix
+        const amount = params.object.paymentMethod?.amount;
+        if (typeof amount !== 'number') {
+            throw new factory.errors.ArgumentNull('object.paymentMethod.amount');
+        }
+
+        const transactionNumber = params.transactionNumber;
+        if (typeof transactionNumber !== 'string' || transactionNumber.length === 0) {
+            throw new factory.errors.ArgumentNull('transactionNumber');
+        }
+
+        await validateSeller(params)(repos);
+
+        let result: IPublishPaymentUrlResult;
+
+        switch (paymentServiceType) {
+            case factory.service.paymentService.PaymentServiceType.CreditCard:
+                const authorizeResult = await CreditCardPayment.authorize(params)(repos);
+                const acsUrl = authorizeResult.execTranResult.acsUrl;
+                if (typeof acsUrl !== 'string' || acsUrl.length === 0) {
+                    throw new factory.errors.ServiceUnavailable('Payment URL unable to publish');
+                }
+                result = {
+                    paymentMethodId: transactionNumber,
+                    paymentUrl: acsUrl
+                };
+
+                break;
+
+            default:
+                throw new factory.errors.NotImplemented(`Payment service '${paymentServiceType}' not implemented`);
+        }
+
+        return result;
+    };
+}
 
 /**
  * 決済方法認証
@@ -104,6 +167,7 @@ export function start(
         product: ProductRepo;
         project: ProjectRepo;
         seller: SellerRepo;
+        serviceOutput: ServiceOutputRepo;
         transaction: TransactionRepo;
         task: TaskRepo;
     }) => {
@@ -195,6 +259,7 @@ function processAuthorizeAccount(
         event: EventRepo;
         project: ProjectRepo;
         seller: SellerRepo;
+        serviceOutput: ServiceOutputRepo;
         transaction: TransactionRepo;
     }): Promise<factory.assetTransaction.pay.ITransaction> => {
         await validateAccount(params)(repos);