+++ src/service/transaction/reserve/factory.ts
@@ -124,13 +124,15 @@ export function createReservedTicket(params: {
     }
 
     const ticketType = createTicketType({ availableOffer: params.availableOffer });
+    const issuedBy = createIssuedBy({ acceptedOffer: params.acceptedOffer, availableOffer: params.availableOffer });
 
     return {
         dateIssued: params.dateIssued,
-        issuedBy: {
-            typeOf: params.event.location.typeOf,
-            name: <string>params.event.location.name?.ja
-        },
+        // issuedBy: {
+        //     typeOf: params.event.location.typeOf,
+        //     name: <string>params.event.location.name?.ja
+        // },
+        issuedBy,
         priceCurrency: factory.priceCurrency.JPY,
         ticketType,
         // totalPrice: ticketOffer.priceSpecification, // いったん不要かと思われる
@@ -190,6 +192,23 @@ function createTicketType(params: {
             }
             : undefined
     };
+}
+
+function createIssuedBy(params: {
+    acceptedOffer: factory.event.screeningEvent.IAcceptedTicketOfferWithoutDetail;
+    availableOffer: factory.offer.IUnitPriceOffer;
+}): factory.reservation.IUnderName<factory.reservationType.EventReservation> {
+    let typeOf = params.acceptedOffer.itemOffered?.serviceOutput?.reservedTicket?.issuedBy?.typeOf;
+    if (typeof typeOf !== 'string') {
+        typeOf = params.availableOffer.project.typeOf;
+    }
+
+    let name = params.acceptedOffer.itemOffered?.serviceOutput?.reservedTicket?.issuedBy?.name;
+    if (typeof name !== 'string') {
+        name = params.availableOffer.project.id;
+    }
+
+    return { name, typeOf };
 }
 
 function validateDate(params: {
@@ -312,11 +331,14 @@ function validateEligibleMembershipType(params: {
             throw new factory.errors.Argument('programMembershipUsed', 'programMembership required');
         }
 
-        const isEligible = eligibleMembershipType.some((membershipType) => membershipType.codeValue === programMembershipUsed.typeOf);
+        // programMembershipUsed.issuedThrough.serviceTypeで検証する
+        const isEligible = eligibleMembershipType.some(
+            (membershipType) => membershipType.codeValue === programMembershipUsed.issuedThrough?.serviceType?.codeValue
+        );
         if (!isEligible) {
             throw new factory.errors.Argument(
                 'programMembershipUsed',
-                `${programMembershipUsed.typeOf} is not eligible for the offer ${params.availableOffer.id}`
+                `${programMembershipUsed.identifier} is not eligible for the offer ${params.availableOffer.id}`
             );
         }
     }
@@ -353,6 +375,7 @@ export function createAdditionalTicketText(params: {
 
 export type IUnitPriceSpecification = factory.priceSpecification.IPriceSpecification<factory.priceSpecificationType.UnitPriceSpecification>;
 
+// tslint:disable-next-line:max-func-body-length
 export function createReservation(params: {
     project: factory.project.IProject;
     id: string;
@@ -412,6 +435,23 @@ export function createReservation(params: {
         project: params.project,
         typeOf: factory.reservationType.EventReservation,
         id: params.id,
+        issuedThrough: {
+            typeOf: factory.product.ProductType.EventService,
+            // issuedThrough.serviceTypeを連携
+            ...(typeof params.reservationFor.offers?.itemOffered.serviceType?.codeValue === 'string')
+                ? {
+                    serviceType: {
+                        codeValue: params.reservationFor.offers.itemOffered.serviceType.codeValue,
+                        inCodeSet: {
+                            typeOf: 'CategoryCodeSet',
+                            identifier: factory.categoryCode.CategorySetIdentifier.ServiceType
+                        },
+                        project: params.project,
+                        typeOf: 'CategoryCode'
+                    }
+                }
+                : undefined
+        },
         additionalProperty: params.additionalProperty,
         bookingTime: params.reserveDate,
         modifiedTime: params.reserveDate,
@@ -425,7 +465,24 @@ export function createReservation(params: {
             ]
         },
         priceCurrency: factory.priceCurrency.JPY,
-        reservationFor: params.reservationFor,
+        // reservationForを最適化
+        reservationFor: {
+            endDate: params.reservationFor.endDate,
+            eventStatus: params.reservationFor.eventStatus,
+            id: params.reservationFor.id,
+            location: params.reservationFor.location,
+            name: params.reservationFor.name,
+            project: params.reservationFor.project,
+            startDate: params.reservationFor.startDate,
+            superEvent: params.reservationFor.superEvent,
+            typeOf: params.reservationFor.typeOf,
+            ...(typeof params.reservationFor.workPerformed?.typeOf === 'string')
+                ? { workPerformed: params.reservationFor.workPerformed }
+                : undefined,
+            ...(params.reservationFor.doorTime instanceof Date)
+                ? { doorTime: params.reservationFor.doorTime }
+                : undefined
+        },
         reservationNumber: params.reservationNumber,
         reservationStatus: factory.reservationStatusType.ReservationPending,
         reservedTicket: params.reservedTicket,